feat: implement AMM liquidity pools contract

+++ contracts/shared/src/lib.rs
@@ -14,6 +14,6 @@ pub use utils::*;
 
 pub fn calculate_percentage(amount: i128, percentage: u32, total_percentage: u32) -> i128 {
     // Calculate using i128 to avoid precision loss
-    let numerator = amount as i128 * percentage as i128;
+    let numerator = amount * percentage as i128;
     numerator / total_percentage as i128
 }
